do not export POSIXLY_CORRECT forever in getopt.m4
Previous code tested whether POSIXLY_CORRECT was set, but not whether it
was exported.  In some situations, e.g., /bin/sh invoked on Linux, the
variable can be set but not exported.  In this case, adding 'export
POSIXLY_CORRECT' causes later tests, which expect to be run with the
existing value of POSIXLY_CORRECT, to fail.

This patch fixes the problem by setting POSIXLY_CORRECT in the body of
the test program itself.